User can help solve the puzzle
This version allows the user to add values to the grid if the program has run out of possibilities. Each time the user adds a new value to the grid, the program tries to complete the new version of the puzzle. It is assumed that the user is adding in correct values.