Changes 3/13/2022
User can input an unsolved sudoku puzzle as an argument in the main method. PuzzleIn.txt is an example of what the file should look like. The program tries to complete the puzzle. If the program runs out of possibilities the user can add values into the console to complete the puzzle. The program prints out the most completed version of the puzzle.