Completed puzzle with recursive backtracking
Updated solve algorithm with recursive backtracking. Iterates through the entire board-- rows then columns. Plugs in a number 1-9 to see if there is a solution. If there is not a valid solution, backtracks a position. At the end, the board is shown in the console along with a message about completion of the puzzle. 
I also made a change to the isValid method. There was a bug with determining if the column contained the value. 
Made another update where user can't manipulate the board. The board is either completed or incomplete after running one time. 
Lastly, restructured program. Cleaned up existing methods and removed others that were not needed.